refactor: extract out blurIImage to reuse in video and image
